[WIP] added reducers and created store to handle initial state
