+++ danceCritiqueForm.js
@@ -1,13 +1,28 @@
 import React from 'react'
 import { reduxForm } from 'redux-form'
+import { StyleSheet, Text, View } from 'react-native'
 
 const DanceCritiqueForm = props => {
   const { handleSubmit } = props
-  return <form onSubmit={handleSubmit}>{/* form body*/}</form>
+  return (
+		<View style={styles.container}>
+			<Text>DanceFest!</Text>
+				{/* form body*/}
+		</View>
+	)
 }
 
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+
 DanceCritiqueForm = reduxForm({
-  // a unique name for the form
   form: 'danceCritique'
 })(DanceCritiqueForm)
