+++ rootReducer.js
@@ -2,9 +2,7 @@ import { combineReducers } from 'redux'
 import { reducer as formReducer } from 'redux-form'
 
 const reducers = {
-  // ...your other reducers here
-  // you have to pass formReducer under 'form' key,
-  // for custom keys look up the docs for 'getFormState'
+	//place other reducers here
   form: formReducer
 }
